Tugas Akhir Web Statistika (Streamlit)/requirements.txt
pandas
streamlit
numpy
scipy.stats
matplotlib
pip
